+++ app/Providers/AppServiceProvider.php
@@ -3,10 +3,12 @@
 namespace App\Providers;
 
 use App\Contracts\ActivityLoggerInterface;
+use App\Contracts\AvatarServiceInterface;
 use App\Contracts\LicenseServiceInterface;
 use App\Models\User;
 use App\Observers\UserObserver;
 use App\Services\ActivityLogger as ActivityLoggerService;
+use App\Services\AvatarService;
 use App\Services\LicenseService;
 use App\Translation\DatabaseLoader;
 use Illuminate\Support\Facades\Cache;
@@ -25,6 +27,7 @@ class AppServiceProvider extends ServiceProvider
 
         $this->app->bind(LicenseServiceInterface::class, LicenseService::class);
         $this->app->bind(ActivityLoggerInterface::class, ActivityLoggerService::class);
+        $this->app->bind(AvatarServiceInterface::class, AvatarService::class);
     }
 
     /**